Common, Agent: Extract credential encoding related methods
