+++ monkey/common/credentials/__init__.py
@@ -3,5 +3,6 @@ from .nt_hash import NTHash
 from .password import Password
 from .ssh_keypair import SSHKeypair
 from .username import Username
+from .encoding import get_plaintext, SecretEncodingConfig
 
 from .credentials import Credentials
