+++ monkey/infection_monkey/exploit/mssqlexec.py
@@ -6,7 +6,7 @@ from typing import Sequence, Tuple
 import pymssql
 
 from common.common_consts.timeouts import LONG_REQUEST_TIMEOUT
-from common.credentials.credentials import get_plaintext
+from common.credentials import get_plaintext
 from common.utils.exceptions import FailedExploitationError
 from infection_monkey.exploit.HostExploiter import HostExploiter
 from infection_monkey.exploit.tools.helpers import get_agent_dst_path
